requirements.txt
telethon==1.34.0
PyYAML==6.0.1
pytest==7.4.4
